+++ src/features/mod.rs
@@ -2,3 +2,4 @@ pub mod add;
 pub mod clear;
 pub mod list;
 pub mod misc;
+pub mod update;